Make the provider almost work on localized Windows
Most of the outstanding issues have been fixed. Unfortunately lots of
indentation fixes are still mixed in. This commit was made primarily to allow my
fixes to be cherry-picked to the original module, should it get a new maintainer
at some point. Imho this module seems like an overkill for editing a clearly
structured text file (Registry.pol).

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>